src/resources/dataIntegration/requirements.txt
schedule
requests
pycryptodome
tqdm
